+++ admin/partials/function/auxiliary/census-single.php
@@ -30,8 +30,8 @@ if (!class_exists('MaBox_Census_Single')) {
 
             add_submenu_page(
                 'index.php',
-                __('发文统计', 'magick-toolbox'),
-                __('发文统计', 'magick-toolbox'),
+                __('发文统计', 'npcink-site-toolbox'),
+                __('发文统计', 'npcink-site-toolbox'),
                 'administrator',
                 'magick-census-single',
                 array(__CLASS__, 'load_content')